cmake: Fix missing `find_package(Doxygen)`
When installing ITK from source (sdist), the CMakeLists.txt
file provided at the root of the project is used. This CMakeLists.txt file
uses the CMake variable `DOXYGEN_EXECUTABLE` assuming that it has been set
,which is the case when compiling ITKPythonPackage using the scripts available
in the `scripts` subdirectory. However, when building ITK from source, this
value was not set and compilation was failing with the following error message:

  `Could NOT find Doxygen (missing: DOXYGEN_EXECUTABLE)`