COMP: Updates for linux aarch64 builds
We need to upgracde almalinux-release to get new GPG keys so we can
install the sudo and ninja-build packages.

We need to update the cmake version for better Python Stable ABI support
to build the 3.11 wheels.

We need to upgrade pip in the image so it knows how to install the
Stable ABI wheels.

We no-longer remove the old wheels after running auditwheel because they
start with the same name.